Refactoring: Add caching controller
  * Moved cache checks into their own controller
  * Added cache checks specs
  * Rubocop controllers
