+++ app/lib/easy_monitor/util/connectors/memcached_connector.rb
@@ -3,7 +3,6 @@ module EasyMonitor
   module Util
     module Connectors
       class MemcachedConnector
-
         attr_accessor :cache
 
         def initialize(cache)
@@ -11,18 +10,17 @@ module EasyMonitor
         end
 
         def memcached_alive?
-          return check_memcached
+          check_memcached
         end
 
         private
 
         def check_memcached
-          return false unless self.cache.write('ping', 'pong')
-          self.cache.read('ping') == 'pong'
+          return false unless cache.write('ping', 'pong')
+
+          cache.read('ping') == 'pong'
         end
       end
     end
   end
 end
-
-
